Updates for py3 and current API
Temporarily commenting out parts that are not working.
Added bytes decoding for when the http response returns the "page not
found" instead of json.
Ran py2to3 to update things like changing "unicode()" to "str()".
Fixed paths for recent media in client.py.
Fixed dict error for 'data' in ['comments'] in models.py.